+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **BREAKING**: Don't merge framework configuration.
 
 ### Removed
+- **BREAKING**: Middleware Authenticate.
 - **BREAKING**: Middleware ValidateCsrfToken.
 
 ## [0.27.0] - 2024-05-31